challs_finished/Shugo_No_Michis_System/shugos/challenge/app/logic_tracker/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
greenlet==3.0.3
